+++ CHANGELOG.md
@@ -33,6 +33,7 @@
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityTemplateDataInfoController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityTemplateDriverInfoController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityController。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.PoacController。
 
 - 为 dubbo 的消费者增加了分组设置。
   - snowflake。